refactor: move satusehat components to be in app module
